Fix the static issues of umediaserver component
:Release Notes:
Address the static coverity issue for umediaserver

:Detailed Notes:
Added the code to resolve parse warnings for
umediaserver.

:Testing Performed:
QWTC-474,QWTC-475,QWTC-476,QWTC-477,QWTC-492,QWTC-496

:QA Notes:
NA

:Issues Addressed:
[WRP-14518] Resolve static analysis in uMS/OSE

+++ src/media_client/uMediaClient.cpp
@@ -520,10 +520,10 @@ bool uMediaClient::stateChange(UMSConnectorHandle* handle, UMSConnectorMessage* 
 
   else if (name == "activeRegion") {
     rect_t active_rc {
-      unmarshalllong(value["x"]),
-      unmarshalllong(value["y"]),
-      unmarshalllong(value["width"]),
-      unmarshalllong(value["height"])
+      (int)unmarshalllong(value["x"]),
+      (int)unmarshalllong(value["y"]),
+      (int)unmarshalllong(value["width"]),
+      (int)unmarshalllong(value["height"])
     };
     return onActiveRegion(active_rc);
   }
